fix(automatically): resolve macOS compilation errors for core-graphics 0.24
- Remove unused CGEventFlags imports in keyboard.rs
- Remove unused CGDisplay import in mouse.rs
- Replace CGEvent::new_scroll_event with CGEvent::new + scroll fields
- Replace CGImage::data_provider with screencapture CLI in screen.rs
- Add std::os::unix::fs::PermissionsExt import in app_index.rs
- Fix unused variable min_version -> _min_version
- Fix u64 to u32 type mismatch in window.rs get_screen_size_macos

+++ src-tauri/crates/claw-automatically/src/platform/window.rs
@@ -704,8 +704,8 @@ fn get_screen_size_macos() -> Result<(u32, u32)> {
 
     let display_id = CGMainDisplayID();
     let display = CGDisplay::new(display_id);
-    let width = display.pixels_wide();
-    let height = display.pixels_high();
+    let width = display.pixels_wide() as u32;
+    let height = display.pixels_high() as u32;
     Ok((width, height))
 }
